+++ src/core_modules/capture-core/components/WidgetEventEdit/EditEventDataEntry/EditEventDataEntry.container.js
@@ -92,7 +92,9 @@ const mapDispatchToProps = (dispatch: ReduxDispatch, props): any => ({
         dispatch(startAsyncUpdateFieldForEditEvent(innerAction, onAsyncUpdateSuccess, onAsyncUpdateError));
     },
     onSave: (orgUnit: OrgUnit) => (eventId: string, dataEntryId: string, formFoundation: RenderFoundation) => {
+        const { onSaveExternal } = props;
         window.scrollTo(0, 0);
+        onSaveExternal && onSaveExternal();
         dispatch(requestSaveEditEventDataEntry(eventId, dataEntryId, formFoundation, orgUnit));
     },
     onSaveAndCompleteEnrollment: (orgUnit: OrgUnit) => (