+++ src/core_modules/capture-core/components/WidgetProfile/OverflowMenu/OverflowMenu.container.js
@@ -1,19 +1,30 @@
 // @flow
 import React from 'react';
 import type { Props } from './OverflowMenu.types';
-import { OverflowMenuComponet } from './OverflowMenu.component';
+import { OverflowMenuComponent } from './OverflowMenu.component';
 import { useAuthorities } from './hooks';
 
-export const OverflowMenu = ({ trackedEntityTypeName, canWriteData, trackedEntity, onDeleteSuccess }: Props) => {
+export const OverflowMenu = ({
+    trackedEntityTypeName,
+    canWriteData,
+    trackedEntity,
+    onDeleteSuccess,
+    displayChangelog,
+    teiId,
+    programAPI,
+}: Props) => {
     const { canCascadeDeleteTei } = useAuthorities();
 
     return (
-        <OverflowMenuComponet
+        <OverflowMenuComponent
             trackedEntityTypeName={trackedEntityTypeName}
             canWriteData={canWriteData}
             canCascadeDeleteTei={canCascadeDeleteTei}
             trackedEntity={trackedEntity}
             onDeleteSuccess={onDeleteSuccess}
+            displayChangelog={displayChangelog}
+            teiId={teiId}
+            programAPI={programAPI}
         />
     );
 };